refactor: useLocalStorage 위치 수정
